+++ src/main/java/net/javaguides/springboot/model/Proveedor.java
@@ -10,7 +10,7 @@ import javax.persistence.Table;
 public class Proveedor {
 	
 	@Id
-	private int nit;
+	private long nit;
 	@Column(name = "nombre_tienda")
 	private String nombre_tienda;
 	@Column(name = "direccion")
@@ -21,17 +21,17 @@ public class Proveedor {
 	public Proveedor() {
 		super();
 	}
-	public Proveedor(int nit, String nombre_tienda, String direccion, int telefono) {
+	public Proveedor(long nit, String nombre_tienda, String direccion, int telefono) {
 		super();
 		this.nit = nit;
 		this.nombre_tienda = nombre_tienda;
 		this.direccion = direccion;
 		this.telefono = telefono;
 	}
-	public int getNit() {
+	public long getNit() {
 		return nit;
 	}
-	public void setNit(int nit) {
+	public void setNit(long nit) {
 		this.nit = nit;
 	}
 	public String getNombre_tienda() {
